@profclems 2020-08-19
```
goos: linux
goarch: amd64
pkg: glab/internal/config
BenchmarkGetEnv-4       83589975                14.0 ns/op
BenchmarkGetEnv-4       83691345                14.7 ns/op
BenchmarkGetEnv-4       85224246                14.0 ns/op
BenchmarkGetEnv-4       84176131                14.0 ns/op
BenchmarkGetEnv-4       78762896                16.2 ns/op
```
This is a great performance improvement. It is very negligible